test(e2e): boot contacts behind a seeded ledger sync trustchain
Contacts gates every mutation on a Ledger Sync status of exactly "ready", so the
suite opened behind a "Turn on Ledger Sync to save contacts" sheet with no
read-only path around it. Seed a trustchain with the CLI and hand it to the app
through the userdata, mirroring initPreSeeded in the ledgerSync suite: the app
boots already a member and the sheet never opens.

Speculos is now needed at seed time to create the trustchain, so the suite is no
longer device-free and its header comment said otherwise. Drop @LNS and gate on
describeIfNotNanoS to match the ledgerSync specs, since Ledger Sync is not
available on Nano S.

The seed setup, trustchain teardown, STAGING guard and llmWalletSync flag would
otherwise sit in two byte-identical copies, so they move to a shared helper both
suites import.
